DM-9943 some fixings based on the review and update the left margin (the side with y axis title) based on the tick string length dynamically.
