+++ src/firefly/js/charts/chartTypes/HistogramTblView.jsx
@@ -11,12 +11,9 @@ import {HelpIcon} from '../../ui/HelpIcon.jsx';
 import {ToolbarButton} from '../../ui/ToolbarButton.jsx';
 
 import * as ChartsCntlr from '../ChartsCntlr.js';
-
 import {HistogramOptions} from '../ui/HistogramOptions.jsx';
 import {Histogram} from '../ui/Histogram.jsx';
-import {HistogramPlotly} from '../ui/HistogramPlotly.jsx';
 import {getChartProperties, updateOnStoreChange, FilterEditorWrapper} from './TblView.jsx';
-import {getAppOptions} from '../../core/AppDataCntlr.js';
 
 import OUTLINE_EXPAND from 'html/images/icons-2014/24x24_ExpandArrowsWhiteOutline.png';
 import SETTINGS from 'html/images/icons-2014/24x24_GearsNEW.png';
@@ -36,12 +33,11 @@ export const HISTOGRAM_TBLVIEW = {
 
 
 function Chart(props) {
-    const {chartId, tblId, chartData, widthPx, heightPx} = props;
+    const {chartId, tblId, chartData, widthPx, heightPx, eventCallback} = props;
     if (!TblUtil.isFullyLoaded(tblId) || !chartData || !heightPx || !widthPx) {
         return (<div/>);
     }
     var { isDataReady, data:histogramData, options:histogramParams} = ChartsCntlr.getChartDataElement(chartId);
-    const HistogramInstance = get(getAppOptions(), 'charts.chartEngine') !== 'plotly' ? Histogram : HistogramPlotly;
     const unit = get(TblUtil.getColumn(TblUtil.getTblById(tblId), histogramParams.columnOrExpr), 'units');
 
     if (isDataReady) {
@@ -62,15 +58,18 @@ function Chart(props) {
         }
 
         return (
-            <HistogramInstance data={histogramData}
-                               desc={histogramParams.columnOrExpr}
-                               binColor='#8c8c8c'
-                               height={heightPx}
-                               width={widthPx}
-                               logs={logs}
-                               xAxis={xAxis}
-                               yAxis={yAxis}
-                               xUnit={unit}
+            <Histogram data={histogramData}
+                       desc={histogramParams.columnOrExpr}
+                       binColor='#8c8c8c'
+                       height={heightPx}
+                       width={widthPx}
+                       logs={logs}
+                       xAxis={xAxis}
+                       yAxis={yAxis}
+                       xUnit={unit}
+                       eventCallback={eventCallback}
+                       chartId={chartId}
+
             />
         );
     } else {
@@ -87,7 +86,8 @@ Chart.propTypes = {
     chartData : PropTypes.object,
     tblId : PropTypes.string,
     widthPx : PropTypes.number,
-    heightPx : PropTypes.number
+    heightPx : PropTypes.number,
+    eventCallback: PropTypes.object
 };
 
 function Options({chartId, optionsKey}) {
